general refactor for logic
